fix(teams-meetings): harden live browser validation
