+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -158,7 +158,7 @@ a5f59c9acbcaa3f82247bf806eb5ba08032373fb853719f0ec9457690f16fc70  module/media-m
 7a5a1de743fff9d5a86515b501ffd88e7fa49970c9f46a5153816e2ebaef7bb0  module/media-store
 c5e3eb1a584f4b8126d9d6c177a840ec9103671e8d1242634ee67db9b5b9e573  module/media-understanding
 c0ffaed532578cf33493992e1ff806b2268b8e3774a92edbaede5cf5bda162a6  module/media-understanding-runtime
-2ae721366251d93b23b42bb349951a62525ec5505696d91dc3b801000e0a0282  module/meeting-runtime
+6d83ea52b4f7406af7e992f31f4b511577bec6990edbe74efe50cfa45703f040  module/meeting-runtime
 bdcdb7db74523c557b6bcaa98a47210ad90f16bc97f9143227d03108c7fd8c23  module/memory-core
 6ae3c62fd51b36439cd3f22286f2e52c6bb27bc85de648650b805c3b7603a487  module/memory-core-engine-runtime
 2bc363010e81c6990c30fd6743e3ffa37320ec39b91abfc222302dc812d8bd33  module/memory-core-host-embedding-registry
