+++ extensions/teams-meetings/src/runtime-probes.test.ts
@@ -1,6 +1,10 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { resolveTeamsMeetingsConfig } from "./config.js";
-import { testTeamsMeetingSpeech, type TeamsMeetingsProbeContext } from "./runtime-probes.js";
+import {
+  testTeamsMeetingListening,
+  testTeamsMeetingSpeech,
+  type TeamsMeetingsProbeContext,
+} from "./runtime-probes.js";
 import type { TeamsMeetingsSession } from "./transports/types.js";
 
 const URL = "https://teams.microsoft.com/l/meetup-join/19%3ameeting_probe%40thread.v2/0";
@@ -43,5 +47,80 @@ describe("Microsoft Teams meeting runtime probes", () => {
     expect(result.speechOutputTimedOut).toBe(true);
     expect(refreshHealth).toHaveBeenCalledTimes(2);
     expect(Date.now()).toBe(200);
+  });
+
+  it("bounds a blocked caption refresh by the per-request listening timeout", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(0);
+    const session = {
+      agentId: "main",
+      chrome: { health: { inCall: true }, launched: true },
+      id: "teams-listen-1",
+      mode: "transcribe",
+      transport: "chrome",
+    } as TeamsMeetingsSession;
+    const refreshCaptionHealth = vi.fn(() => new Promise<void>(() => {}));
+    const context = {
+      config: resolveTeamsMeetingsConfig({ chrome: { joinTimeoutMs: 30_000 } }),
+      hasHealthHandle: () => true,
+      isReusable: () => false,
+      join: vi.fn(async () => ({ session, spoken: false })),
+      list: () => [],
+      refreshCaptionHealth,
+      refreshHealth: () => {},
+      resolveAgentId: () => "main",
+    } satisfies TeamsMeetingsProbeContext;
+
+    const pending = testTeamsMeetingListening(context, {
+      mode: "transcribe",
+      timeoutMs: 300,
+      url: URL,
+    });
+    await vi.advanceTimersByTimeAsync(350);
+    const result = await pending;
+
+    expect(result.listenTimedOut).toBe(true);
+    expect(refreshCaptionHealth).toHaveBeenCalledTimes(1);
+    expect(Date.now()).toBe(350);
+  });
+
+  it("refreshes captions before a short listening timeout expires", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(0);
+    const session = {
+      agentId: "main",
+      chrome: { health: { inCall: true }, launched: true },
+      id: "teams-listen-2",
+      mode: "transcribe",
+      transport: "chrome",
+    } as TeamsMeetingsSession;
+    const refreshCaptionHealth = vi.fn(async () => {
+      session.chrome!.health = {
+        ...session.chrome!.health,
+        lastCaptionText: "Caption already waiting",
+        transcriptLines: 1,
+      };
+    });
+    const context = {
+      config: resolveTeamsMeetingsConfig({ chrome: { joinTimeoutMs: 30_000 } }),
+      hasHealthHandle: () => true,
+      isReusable: () => false,
+      join: vi.fn(async () => ({ session, spoken: false })),
+      list: () => [],
+      refreshCaptionHealth,
+      refreshHealth: () => {},
+      resolveAgentId: () => "main",
+    } satisfies TeamsMeetingsProbeContext;
+
+    const result = await testTeamsMeetingListening(context, {
+      mode: "transcribe",
+      timeoutMs: 100,
+      url: URL,
+    });
+
+    expect(result.listenVerified).toBe(true);
+    expect(refreshCaptionHealth).toHaveBeenCalledTimes(1);
+    expect(Date.now()).toBe(0);
+    expect(vi.getTimerCount()).toBe(0);
   });
 });
